requirements.txt
uuid
numpy
kokoro
soundfile
fastapi[standard]
loguru
chatterbox-tts
faster_whisper
torchaudio
